+++ src/simulation/resources.rs
@@ -2,7 +2,7 @@ use bevy::prelude::*;
 
 #[derive(Resource)]
 pub struct SimulationSettings {
-    pub size: f32,
+    pub size: usize,
     pub max_particle_radius: f32,
     pub gravity: GravityMode,
 }
@@ -10,7 +10,7 @@ pub struct SimulationSettings {
 impl Default for SimulationSettings {
     fn default() -> Self {
         Self {
-            size: 1024.0,
+            size: 1024,
             max_particle_radius: 2.0,
             // gravity: GravityMode::Constant(Vec2::new(0.0, -9.8)),
             gravity: GravityMode::Point(Vec2::ZERO, 24000.0),
